Weight Tolerance Parameter
Weight Tolerance handles how to combine weights of edges that have been matched. If weight tolerance = 0, weights are summed If weight tolerance = 1. the weights are summed if the edges are the same distance to start or accept node, otherwise is averaged. If weight tolerance = 2, the weights are summed if their edges are next to the same parts, otherwise they are averaged.